requirements.txt
Flask[async]
requests
aiohttp
googleapis-common-protos
pycryptodome
protobuf
Werkzeug
firebase-admin
APScheduler
Flask-Limiter
python-dotenv
pytest
pytest-asyncio